# scray-examples/querying/rest-mongo/swagger-spec.json
{
  "swagger": "2.0",
  "info": {
    "description": "Simple API to query MongoDB data",
    "version": "1.0.0",
    "title": "Swagger Server",
    "termsOfService": "",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost",
  "basePath": "/scray/examples/1.0.0",
  "tags": [
    {
      "name": "developers"
    },
    {
      "name": "admins"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/mongodb": {
      "get": {
        "tags": [
          "developers"
        ],
        "summary": "get query result",
        "description": "By passing a querry, you can  ",
        "operationId": "getQueryResult",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "queryId",
            "in": "query",
            "description": "pass the name of the query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "search results matching criteria",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/QueryResult"
              }
            }
          },
          "400": {
            "description": "bad input parameter"
          }
        }
      },
      "post": {
        "tags": [
          "admins"
        ],
        "summary": "adds an query",
        "description": "Adds new query",
        "operationId": "addQuery",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "MongoDB query",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Query"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "item created"
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        }
      }
    }
  },
  "definitions": {
    "QueryResult": {
      "type": "object",
      "required": [
        "id",
        "value"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid",
          "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name": {
          "type": "string",
          "example": "Super KPI"
        },
        "value": {
          "type": "string",
          "example": "42"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "Query": {
      "type": "object",
      "required": [
        "collectionName",
        "databaseName",
        "id",
        "query",
        "serverAddress"
      ],
      "properties": {
        "id": {
          "type": "string",
          "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "serverAddress": {
          "type": "string",
          "example": "127.0.0.1:27017"
        },
        "databaseName": {
          "type": "string",
          "example": "db1"
        },
        "collectionName": {
          "type": "string",
          "example": "col1"
        },
        "query": {
          "type": "string",
          "example": "SELECT * FROM inventory WHERE status = \"D\""
        },
        "description": {
          "type": "string",
          "example": "Super description"
        }
      }
    }
  }
}